+++ lib/src/cache/cache_proxy.dart
@@ -1,4 +1,7 @@
-import './options.dart';
+import 'package:gql_exec/gql_exec.dart';
+import 'package:gql/ast.dart';
+import 'package:meta/meta.dart';
+
 import './cache.dart';
 
 class CacheProxy {
@@ -15,22 +18,53 @@ class CacheProxy {
         _queryId = queryId;
 
   Map<String, dynamic> readQuery(
-    ReadQueryOptions options,
-  ) =>
-      _cache.readQuery(options.rebuild((b) => b..optimistic ??= false));
+    Request request, {
+    bool optimistic,
+  }) =>
+      _cache.readQuery(request, optimistic: optimistic ?? false);
 
-  Map<String, dynamic> readFragment(
-    ReadFragmentOptions options,
-  ) =>
-      _cache.readFragment(options.rebuild((b) => b..optimistic ??= false));
+  Map<String, dynamic> readFragment({
+    @required DocumentNode fragment,
+    @required Map<String, dynamic> idFields,
+    String fragmentName,
+    Map<String, dynamic> variables,
+    bool optimistic,
+  }) =>
+      _cache.readFragment(
+        fragment: fragment,
+        idFields: idFields,
+        fragmentName: fragmentName,
+        variables: variables,
+        optimistic: optimistic ?? false,
+      );
 
   void writeQuery(
-    WriteQueryOptions options,
-  ) =>
-      _cache.writeQuery(options, _optimistic, _queryId);
+    Request request,
+    Map<String, dynamic> data, {
+    String queryId,
+  }) =>
+      _cache.writeQuery(
+        request,
+        data,
+        optimistic: _optimistic,
+        queryId: queryId,
+      );
 
-  void writeFragment(
-    WriteFragmentOptions options,
-  ) =>
-      _cache.writeFragment(options, _optimistic, _queryId);
+  void writeFragment({
+    @required DocumentNode fragment,
+    @required Map<String, dynamic> idFields,
+    @required Map<String, dynamic> data,
+    String fragmentName,
+    Map<String, dynamic> variables,
+    String queryId,
+  }) =>
+      _cache.writeFragment(
+        fragment: fragment,
+        idFields: idFields,
+        data: data,
+        fragmentName: fragmentName,
+        variables: variables,
+        optimistic: _optimistic,
+        queryId: queryId,
+      );
 }